backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfminer.six
spacy
